requirements.txt
mysql-connector-python==8.2.0
pandas==2.2.0